remove some duplication by moving shared helpers to cache utils.
