+++ services/ui_backend_service/data/cache/utils.py
@@ -7,6 +7,7 @@ from urllib.parse import urlparse
 from botocore.exceptions import ClientError, NoCredentialsError
 from services.ui_backend_service.features import FEATURE_S3_DISABLE
 
+# Generic helpers
 
 def batchiter(it, batch_size):
     it = iter(it)
@@ -23,6 +24,19 @@ def decode(path):
     with GzipFile(path) as f:
         obj = pickle.load(f)
         return obj
+
+# Cache Action helpers
+
+
+def artifact_cache_id(location):
+    "construct a unique cache key for artifact location"
+    return 'search:artifactdata:%s' % location
+
+
+def artifact_location_from_key(x):
+    "extract location from the artifact cache key"
+    return x[len("search:artifactdata:"):]
+
 
 # Cache action stream output helpers
